moves files to directory
made a lib directory for code to go to
